+++ cmd/contact.go
@@ -42,7 +42,7 @@ var contactEmailCmd = &cobra.Command{
 	Use:   "email",
 	Short: "Email a contact.",
 	Args:  cobra.ExactArgs(1),
-	Run:   manage.Email,
+	// Run:   manage.Email,
 }
 
 func init() {
@@ -54,9 +54,30 @@ func init() {
 	contactCmd.AddCommand(contactEditCmd)
 	contactCmd.AddCommand(contactRemoveCmd)
 	contactCmd.AddCommand(contactViewCmd)
-	contactCmd.AddCommand(contactEmailCmd)
+	// contactCmd.AddCommand(contactEmailCmd)
 
-	// // View
+	// Add
+	contactFlags(contactAddCmd)
+	contactAddCmd.Flags().BoolP("fast", "f", false, "create contact with only a name and email flags.")
+
+	// Edit
+	// contactFlags(contactEditCmd)
+
+	// View
 	// contactViewCmd.Flags().StringP("output", "o", "", "save the results to a file")
 	// contactViewCmd.Flags().Bool("force", false, "force the results to be saved to an existing file")
+}
+
+func contactFlags(cmd *cobra.Command) {
+	cmd.Flags().StringP("name", "n", "", "name for this contact")
+	cmd.Flags().StringP("email", "e", "", "email for this contact")
+	cmd.Flags().String("phone", "", "phone number for this contact")
+
+	cmd.Flags().BoolP("contacted", "c", false, "contact was made")
+
+	// cmd.Flags().Bool("lead", false, "contact is a lead")
+	// cmd.Flags().Bool("customer", false, "contact is a customer")
+	// cmd.Flags().Bool("subscriber", false, "contact is a subscriber")
+	// cmd.Flags().Bool("advocate", false, "contact is a advocate")
+	// cmd.Flags().String("other", "", "some other contact relationship")
 }