+++ cmd/root.go
@@ -1,15 +1,12 @@
 package cmd
 
 import (
-	"broadcastle.co/code/crm/code/manage"
-	homedir "github.com/mitchellh/go-homedir"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
-var cfgFile string
-var debugLog bool
+// var cfgFile string
+// var debugLog bool
 
 // RootCmd represents the base command when called without any subcommands
 var RootCmd = &cobra.Command{
@@ -24,44 +21,44 @@ func Execute() {
 		logrus.Fatal(err)
 	}
 
-	if debugLog {
-		logrus.SetLevel(logrus.DebugLevel)
-		logrus.Info("debug text enabled")
-	}
+	// if debugLog {
+	// 	logrus.SetLevel(logrus.DebugLevel)
+	// 	logrus.Info("debug text enabled")
+	// }
 
-	manage.Init()
-	defer manage.Close()
+	// manage.Init()
+	// defer manage.Close()
 }
 
 func init() {
-	cobra.OnInitialize(initConfig)
+	// cobra.OnInitialize(initConfig)
 
-	RootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.crm.yaml)")
-	RootCmd.PersistentFlags().BoolVar(&debugLog, "debug", false, "debug output")
+	// RootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.crm.yaml)")
+	// RootCmd.PersistentFlags().BoolVar(&debugLog, "debug", false, "debug output")
 
 }
 
 // initConfig reads in config file and ENV variables if set.
-func initConfig() {
-	if cfgFile != "" {
-		// Use config file from the flag.
-		viper.SetConfigFile(cfgFile)
-	} else {
-		// Find home directory.
-		home, err := homedir.Dir()
-		if err != nil {
-			logrus.Fatal(err)
-		}
+// func initConfig() {
+// 	if cfgFile != "" {
+// 		// Use config file from the flag.
+// 		viper.SetConfigFile(cfgFile)
+// 	} else {
+// 		// Find home directory.
+// 		home, err := homedir.Dir()
+// 		if err != nil {
+// 			logrus.Fatal(err)
+// 		}
 
-		// Search config in home directory with name ".crm" (without extension).
-		viper.AddConfigPath(home)
-		viper.SetConfigName(".crm")
-	}
+// 		// Search config in home directory with name ".crm" (without extension).
+// 		viper.AddConfigPath(home)
+// 		viper.SetConfigName(".crm")
+// 	}
 
-	viper.AutomaticEnv() // read in environment variables that match
+// 	viper.AutomaticEnv() // read in environment variables that match
 
-	// If a config file is found, read it in.
-	if err := viper.ReadInConfig(); err == nil {
-		logrus.Debugf("Using config file: %s", viper.ConfigFileUsed())
-	}
-}
+// 	// If a config file is found, read it in.
+// 	if err := viper.ReadInConfig(); err == nil {
+// 		logrus.Debugf("Using config file: %s", viper.ConfigFileUsed())
+// 	}
+// }